moves wait_until_404 into testing_utils.py
